requirements.txt
openai==0.27.0
pytube==12.1.2
gradio==3.14.0
uvicorn==0.20.0
fastapi==0.88.0
loguru==0.6.0
pyyaml